refactored tests.gofmt project.
