+++ client/mocks_client_gen_test.go
@@ -4,7 +4,6 @@
 package client
 
 import (
-	
 	protocol "github.com/cosminrentea/gobbler/protocol"
 	gomock "github.com/golang/mock/gomock"
 )
