+++ server/auth/mocks_auth_gen_test.go
@@ -5,7 +5,7 @@ package auth
 
 import (
 	protocol "github.com/cosminrentea/gobbler/protocol"
-	
+
 	gomock "github.com/golang/mock/gomock"
 )
