+++ server/benchmarking_apns_test.go
@@ -4,14 +4,14 @@ import (
 	"bytes"
 	"fmt"
 	log "github.com/Sirupsen/logrus"
-	"github.com/golang/mock/gomock"
-	"github.com/sideshow/apns2"
 	"github.com/cosminrentea/gobbler/client"
 	"github.com/cosminrentea/gobbler/server/apns"
 	"github.com/cosminrentea/gobbler/server/connector"
 	"github.com/cosminrentea/gobbler/server/router"
 	"github.com/cosminrentea/gobbler/server/websocket"
 	"github.com/cosminrentea/gobbler/testutil"
+	"github.com/golang/mock/gomock"
+	"github.com/sideshow/apns2"
 	"github.com/stretchr/testify/assert"
 	"io/ioutil"
 	"net/http"
