+++ server/router/mocks_router_gen_test.go
@@ -8,7 +8,7 @@ import (
 	auth "github.com/cosminrentea/gobbler/server/auth"
 	cluster "github.com/cosminrentea/gobbler/server/cluster"
 	kvstore "github.com/cosminrentea/gobbler/server/kvstore"
-	
+
 	store "github.com/cosminrentea/gobbler/server/store"
 	gomock "github.com/golang/mock/gomock"
 )
