+++ server/sms/sms_gateway.go
@@ -3,13 +3,13 @@ package sms
 import (
 	"context"
 	"encoding/json"
-	"github.com/cosminrentea/gobbler/server/connector"
-	"time"
 	log "github.com/Sirupsen/logrus"
 	"github.com/cosminrentea/gobbler/protocol"
+	"github.com/cosminrentea/gobbler/server/connector"
 	"github.com/cosminrentea/gobbler/server/metrics"
 	"github.com/cosminrentea/gobbler/server/router"
 	"github.com/cosminrentea/gobbler/server/store"
+	"time"
 )
 
 const (
