+++ server/sms/sms_integration_test.go
@@ -2,18 +2,14 @@ package sms
 
 import (
 	"encoding/json"
-	"github.com/cosminrentea/gobbler/protocol"
+	"fmt"
 	"github.com/cosminrentea/gobbler/server/auth"
-	"github.com/cosminrentea/gobbler/server/kvstore"
 	"github.com/cosminrentea/gobbler/server/router"
 	"github.com/cosminrentea/gobbler/server/store/dummystore"
 	"github.com/cosminrentea/gobbler/testutil"
 	"github.com/stretchr/testify/assert"
 	"io/ioutil"
 	"net/http"
-	//"os"
-	"fmt"
-	"math/rand"
 	"os"
 	"testing"
 	"time"
@@ -33,7 +29,7 @@ func Test_HttpClientRecreation(t *testing.T) {
 	msg := encodeProtocolMessage(t, 2)
 
 	err := sender.Send(&msg)
-	a.Equal(ErrRetryFailed,err)
+	a.Equal(ErrRetryFailed, err)
 
 	a.Equal(0, expectedRequestNo, "Three retries should be made by sender.")
 	time.Sleep(timeInterval)
@@ -56,9 +52,9 @@ func Test_NexmoHTTPError(t *testing.T) {
 	msgStore := dummystore.New(kvStore)
 	accessManager := auth.NewAllowAllAccessManager(true)
 
-	router := router.New(accessManager, msgStore, kvStore, nil)
+	unstartedRouter := router.New(accessManager, msgStore, kvStore, nil)
 
-	gw, err := New(router, sender, config)
+	gw, err := New(unstartedRouter, sender, config)
 	a.NoError(err)
 	err = gw.Start()
 	a.NoError(err)
@@ -66,7 +62,7 @@ func Test_NexmoHTTPError(t *testing.T) {
 	msg := encodeProtocolMessage(t, 2)
 	err = gw.route.Deliver(&msg, false)
 	a.NoError(err)
-	time.Sleep(4 * timeInterval)
+	time.Sleep(5 * timeInterval)
 	a.Equal(0, expectedRequestNo, "Three retries should be made by sender.")
 	a.Equal(msg.ID, gw.LastIDSent, "Retry failed.Last id  sent should be msgId")
 
@@ -92,9 +88,9 @@ func Test_NexmoInvalidSenderError(t *testing.T) {
 	msgStore := dummystore.New(kvStore)
 	accessManager := auth.NewAllowAllAccessManager(true)
 
-	router := router.New(accessManager, msgStore, kvStore, nil)
+	unstartedRouter := router.New(accessManager, msgStore, kvStore, nil)
 
-	gw, err := New(router, sender, config)
+	gw, err := New(unstartedRouter, sender, config)
 	a.NoError(err)
 	err = gw.Start()
 	a.NoError(err)
@@ -102,7 +98,6 @@ func Test_NexmoInvalidSenderError(t *testing.T) {
 	msg := encodeProtocolMessage(t, 2)
 	err = gw.route.Deliver(&msg, false)
 	a.NoError(err)
-	//time.Sleep(4 * timeInterval)
 	time.Sleep(timeInterval)
 	a.Equal(0, expectedRequestNo, "Only one try should be made by sender.")
 	a.Equal(msg.ID, gw.LastIDSent, "No Retry needed.Last id  sent should be msgId")
@@ -111,7 +106,6 @@ func Test_NexmoInvalidSenderError(t *testing.T) {
 	time.Sleep(timeInterval)
 	a.NoError(err)
 }
-
 
 func Test_NexmoResponseCodeError(t *testing.T) {
 	defer testutil.EnableDebugForMethod()()
@@ -125,14 +119,14 @@ func Test_NexmoResponseCodeError(t *testing.T) {
 
 	sender := createNexmoSender(t)
 	config := createConfig()
-	kvStore, f := createKVStore(t, "/guble_sms_nexmo_invalid_sender_error")
+	kvStore, f := createKVStore(t, "/guble_sms_nexmo_responde_code_error")
 	defer os.Remove(f)
 	msgStore := dummystore.New(kvStore)
 	accessManager := auth.NewAllowAllAccessManager(true)
 
-	router := router.New(accessManager, msgStore, kvStore, nil)
+	unstartedRouter := router.New(accessManager, msgStore, kvStore, nil)
 
-	gw, err := New(router, sender, config)
+	gw, err := New(unstartedRouter, sender, config)
 	a.NoError(err)
 	err = gw.Start()
 	a.NoError(err)
@@ -149,59 +143,6 @@ func Test_NexmoResponseCodeError(t *testing.T) {
 	a.NoError(err)
 }
 
-
-func createConfig() Config {
-	topic := "/sms"
-	worker := 1
-	intervalMetrics := true
-	return Config{
-		Workers:  &worker,
-		SMSTopic: &topic,
-		Name:     "test_gateway",
-		Schema:   SMSSchema,
-
-		IntervalMetrics: &intervalMetrics,
-	}
-
-}
-
-func createKVStore(t *testing.T, filename string) (kvstore.KVStore, string) {
-	a := assert.New(t)
-	f := tempFilename(filename)
-	//defer os.Remove(f)
-
-	//create a KVStore with sqlite3
-	kvStore := kvstore.NewSqliteKVStore(f, true)
-	err := kvStore.Open()
-	if err != nil {
-		a.FailNow("KVStore could not be opened.")
-	}
-	a.NotNil(kvStore)
-	return kvStore, f
-}
-
-func encodeProtocolMessage(t *testing.T, ID int) protocol.Message {
-	a := assert.New(t)
-	sms := NexmoSms{
-		To:   "toNumber",
-		From: "FromNUmber",
-		Text: "body",
-	}
-	d, err := json.Marshal(&sms)
-	if err != nil {
-		a.FailNow("Could not obtain message")
-	}
-
-	msg := protocol.Message{
-		Path:          protocol.Path(SMSDefaultTopic),
-		UserID:        "samsa",
-		ApplicationID: "sms",
-		ID:            uint64(ID),
-		Body:          d,
-	}
-	return msg
-}
-
 func createNexmoSender(t *testing.T) Sender {
 	a := assert.New(t)
 	nexmoSender, err := NewNexmoSender(KEY, SECRET)
@@ -214,12 +155,8 @@ func createNexmoSender(t *testing.T) Sender {
 func noResponseFromNexmoHandler(t *testing.T, noOfReq *int) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		a := assert.New(t)
-		defer r.Body.Close()
-		body, _ := ioutil.ReadAll(r.Body)
-
-		var sms NexmoSms
-		json.Unmarshal(body, &sms)
-		a.Equal("body", sms.Text)
+		sentSms := decodeSMSMessage(t, r)
+		a.Equal("body", sentSms.Text)
 		*noOfReq--
 	}
 }
@@ -227,22 +164,12 @@ func noResponseFromNexmoHandler(t *testing.T, noOfReq *int) http.HandlerFunc {
 func invalidSenderNexmoHandler(t *testing.T, noOfReq *int) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		a := assert.New(t)
-		defer r.Body.Close()
-		body, _ := ioutil.ReadAll(r.Body)
-
-		var sms NexmoSms
-		json.Unmarshal(body, &sms)
-		a.Equal("body", sms.Text)
+		sentSms := decodeSMSMessage(t, r)
+		a.Equal("body", sentSms.Text)
 		*noOfReq--
 
-		nexmoResponse := composeNexmoMessageResponse(sms, ResponseInvalidSenderAddress)
-
-		jData, err := json.Marshal(nexmoResponse)
-		if err != nil {
-			a.FailNow("Nexmo Response encoding failed.")
-		}
-		w.Header().Set("Content-Type", "application/json")
-		w.Write(jData)
+		nexmoResponse := composeNexmoMessageResponse(sentSms, ResponseInvalidSenderAddress, 1)
+		writeNexmoResponse(nexmoResponse, t, w)
 
 	}
 }
@@ -250,41 +177,15 @@ func invalidSenderNexmoHandler(t *testing.T, noOfReq *int) http.HandlerFunc {
 func responseInternalErrorNexmoHandler(t *testing.T, noOfReq *int) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		a := assert.New(t)
-		defer r.Body.Close()
-		body, _ := ioutil.ReadAll(r.Body)
-
-		var sms NexmoSms
-		json.Unmarshal(body, &sms)
-		a.Equal("body", sms.Text)
+		sentSms := decodeSMSMessage(t, r)
+		a.Equal("body", sentSms.Text)
 		*noOfReq--
 
-		nexmoResponse := composeNexmoMessageResponse(sms, ResponseInternalError)
+		nexmoResponse := composeNexmoMessageResponse(sentSms, ResponseInternalError, 1)
 
-		jData, err := json.Marshal(nexmoResponse)
-		if err != nil {
-			a.FailNow("Nexmo Response encoding failed.")
-		}
-		w.Header().Set("Content-Type", "application/json")
-		w.Write(jData)
+		writeNexmoResponse(nexmoResponse, t, w)
 
 	}
-}
-
-func composeNexmoMessageResponse(sms NexmoSms, code ResponseCode) *NexmoMessageResponse {
-	nexmoResponse := new(NexmoMessageResponse)
-	nexmoResponse.MessageCount = 1
-	response := NexmoMessageReport{
-		Status:           code,
-		MessageID:        "msgID",
-		To:               sms.To,
-		ClientReference:  "",
-		RemainingBalance: "2",
-		MessagePrice:     "0.005",
-		Network:          "TELEKOM",
-		ErrorText:        "",
-	}
-	nexmoResponse.Messages = []NexmoMessageReport{response}
-	return nexmoResponse
 }
 
 func dummyNexmoEndpointWithHandlerFunc(t *testing.T, expectedRequestNo *int, port string, handler func(t *testing.T, no *int) http.HandlerFunc) {
@@ -294,8 +195,23 @@ func dummyNexmoEndpointWithHandlerFunc(t *testing.T, expectedRequestNo *int, por
 	http.ListenAndServe(port, serveMux)
 }
 
-func createRandomPort(min, max int) string {
-	rand.Seed(time.Now().UnixNano())
-	port := rand.Intn(max-min) + min
-	return fmt.Sprintf(":%d", port)
+func decodeSMSMessage(t *testing.T, r *http.Request) NexmoSms {
+	a := assert.New(t)
+	defer r.Body.Close()
+	body, _ := ioutil.ReadAll(r.Body)
+	var sentSms NexmoSms
+	err := json.Unmarshal(body, &sentSms)
+	if err != nil {
+		a.FailNow("Could not decode sender sms.")
+	}
+	return sentSms
+}
+func writeNexmoResponse(nexmoResponse *NexmoMessageResponse, t *testing.T, w http.ResponseWriter) {
+	a := assert.New(t)
+	jData, err := json.Marshal(nexmoResponse)
+	if err != nil {
+		a.FailNow("Nexmo Response encoding failed.")
+	}
+	w.Header().Set("Content-Type", "application/json")
+	w.Write(jData)
 }
